requirements.txt
asgiref==3.4.1
Django==3.2.9
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
environ==1.0
pytz==2021.3
sqlparse==0.4.2
